优化渲染数据管理与骨骼数据处理逻辑
重构了 `OpenGLShaderStorageBuffer` 的 `GetSize` 方法，返回类型从 `uint32_t` 改为 `size_t`，以支持更大的数据范围。

在 `OpenGLUniformBuffer` 中，新增 `Renderer::DataAllocate` 方法，用于高效分配对齐数据，替换了原先的 `Buffer` 拷贝逻辑。

在 `ObjectUniformBuffer` 中优化了骨骼数据处理：
- 使用 `memcpy` 替代循环赋值。
- 增加了 `m_BonesDirty` 标志位，优化上传逻辑以减少不必要的数据传输。

新增渲染命令队列数据池管理：
- 引入 `DATA_POOL_MAX_SIZE`（512MB）限制。
- 新增 `DataAllocate` 方法，支持对齐分配和数据拷贝。
- 增加数据池容量统计功能，并在 `Application.cpp` 中显示。

删除了 `RenderCommand` 类及其相关实现，移除了 `Clear` 方法。
移除了 `Renderer` 类中的 `SubmitSubmesh` 方法及其相关逻辑。

修复了多个文件的 BOM 编码问题，统一了代码格式。

+++ Prism/src/Platform/OpenGL/Buffer/OpenGLShaderStorageBuffer.cpp
@@ -71,7 +71,7 @@ namespace Prism
     {
         return m_RendererID;
     }
-    uint32_t OpenGLShaderStorageBuffer::GetSize() const
+    size_t OpenGLShaderStorageBuffer::GetSize() const
     {
         return m_Size;
     }
